+++ src/c/game_over_window.h
@@ -1,4 +1,4 @@
 #pragma once
 
-void game_over_window_create(bool, int, int);
+void game_over_window_create(bool, uint8_t, uint8_t);
 void game_over_window_destroy();
